add visibility modifiers to structs and fields, support struct attributes
closes #4
supersedes #5

The parser now accepts visibility modifiers on struct declarations and
their fields inside the `crepe!` block.

Also attributes can be added on struct definitions, allowing custom derive
or other attributes to be associated to the types.

The attributes can be placed after an `@input` or `@output` modifier and
as usual before the struct declaration.