make hasher configurable
Benchmarks have been a bit flaky and I suspect that it could have been because of the default hasher used.
This change renames the original `run()` function to `run_with_hasher()` that has a type paremeter for the hasher.
To keep compatbility (and convenience!) the `run()` function is now a small wrapper that uses the default `RandomState` hasher.

The benchmarks have been updated to use `fnv` for hashing, as it's a deterministic hasher and should give more robust results.